+++ packages/ui/src/hooks/useGetEncodedAddress.tsx
@@ -1,6 +1,7 @@
 import { useApi } from '../contexts/ApiContext'
 import { useCallback } from 'react'
 import { encodesubstrateAddress } from '../utils/encodeSubstrateAddress'
+import { u8aToHex, isU8a } from '@polkadot/util'
 
 export const useGetEncodedAddress = () => {
   const { chainInfo } = useApi()
@@ -9,6 +10,13 @@ export const useGetEncodedAddress = () => {
     (address: string | Uint8Array | undefined) => {
       if (!chainInfo || !address) {
         return
+      }
+
+      if (chainInfo?.isEthereum) {
+        const res = isU8a(address)
+          ? u8aToHex(address as Uint8Array).toString()
+          : (address as string)
+        return res.slice(0, 42)
       }
 
       return encodesubstrateAddress(address, chainInfo.ss58Format)
